requirements.txt
Flask == 1.1.2
flask-marshmallow == 0.11.0
Flask-SQLAlchemy == 2.4.1
Flask-WTF == 0.14.3
Flask-Login == 0.5.0
Flask-Mail == 0.9.1
Jinja2 == 2.11.1
python-Levenshtein == 0.12.0
requests == 2.23.0
SQLAlchemy == 1.3.15
urllib3 == 1.25.8
WTForms == 2.2.1